Refactor for readibility
