+++ internal/validator/validator.go
@@ -61,146 +61,6 @@ type configLoaded struct {
 
 type probeCallable func(container.ContainerInterface, *canaryv1.Probe) (bool, error)
 
-type model struct {
-	sub              chan checkResult
-	container        container.ContainerInterface
-	containerStarted bool
-	results          []checkResult
-	allChecksPassed  bool
-	spinner          spinner.Model
-	progress         progress.Model
-	debug            bool
-	image            string
-	validator        *canaryv1.Validator
-	configPath       string
-	err              error
-	tty              bool
-}
-
-func (m model) Init() tea.Cmd {
-	return tea.Batch(
-		spinner.Tick,
-		waitForChecks(m.sub),
-		loadConfig(m.configPath),
-	)
-}
-
-func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
-	switch msg := msg.(type) {
-	case tea.KeyMsg:
-		switch msg.String() {
-		// These keys should exit the program.
-		case "ctrl+c", "q":
-			m.allChecksPassed = false
-			if m.containerStarted {
-				return m, shutdown(m.container)
-			} else {
-				return m, tea.Quit
-			}
-		default:
-			return m, nil
-		}
-
-	case configLoaded:
-		if msg.Error != nil {
-			m.err = msg.Error
-			return m, tea.Batch(tea.Printf("Error: %s\n", msg.Error.Error()), tea.Quit)
-		}
-		m.validator = msg.Config
-		return m, startContainer(m.image, m.validator)
-
-	case containerStarted:
-		m.containerStarted = true
-		m.container = msg.Container
-		var checks []tea.Cmd
-		if m.debug {
-			status, _ := m.container.Status()
-			checks = append(checks, tea.Printf("Running container with command '%s'", status.RunCommand))
-		}
-		checks = append(checks, tea.Printf("Validating %s against %s", highlightStyle(m.image), highlightStyle(m.validator.Name)))
-		for _, check := range m.validator.Checks {
-			checks = append(checks, runCheck(m.sub, m.container, check))
-		}
-		return m, tea.Batch(checks...)
-
-	case containerStopped:
-		return m, tea.Quit
-
-	case containerFailed:
-		m.err = msg.Error
-		return m, tea.Batch(tea.Printf("Error: %s\n", m.err.Error()), tea.Quit)
-
-	case checkResult:
-		m.results = append(m.results, msg)
-		if !msg.Passed {
-			m.allChecksPassed = false
-		}
-		if len(m.results) == len(m.validator.Checks) {
-			var commands []tea.Cmd
-
-			if m.allChecksPassed {
-				commands = append(commands, tea.Println(passedStyle("validation passed")))
-			} else {
-				commands = append(commands, tea.Println(failedStyle("validation failed")))
-			}
-			if !m.allChecksPassed && m.debug {
-				commands = append(commands, tea.Println("Leaving container running for debugging..."))
-			} else {
-				commands = append(commands, shutdown(m.container))
-			}
-			return m, tea.Batch(commands...)
-		}
-		if m.allChecksPassed {
-			m.progress.FullColor = "10"
-		} else {
-			m.progress.FullColor = "9"
-		}
-		cmd := m.progress.SetPercent(float64(len(m.results)) / float64(len(m.validator.Checks)))
-		return m, tea.Batch(
-			tea.Printf(" %-50s [%s]", msg.Description, getStatus(msg.Passed, msg.Error)),
-			waitForChecks(m.sub), cmd)
-
-	case spinner.TickMsg:
-		var cmd tea.Cmd
-		m.spinner, cmd = m.spinner.Update(msg)
-		return m, cmd
-
-	case progress.FrameMsg:
-		progressModel, cmd := m.progress.Update(msg)
-		m.progress = progressModel.(progress.Model)
-		return m, cmd
-
-	default:
-		return m, nil
-	}
-}
-
-func (m model) View() string {
-	var s string
-
-	if m.validator == nil {
-		s += fmt.Sprintf("%s Loading config\n", m.spinner.View())
-	} else {
-		if !m.containerStarted {
-			if m.tty {
-				s += fmt.Sprintf("%s Starting container\n", m.spinner.View())
-			} else {
-				s += "Starting container\n"
-			}
-		} else {
-			if m.tty && len(m.results) < len(m.validator.Checks) {
-				s += m.progress.View() + "\n"
-			}
-		}
-	}
-	s += helpStyle("Press q to quit...")
-	return s
-}
-
-func (m model) Passed() bool {
-	return m.allChecksPassed
-}
-
 func Validate(image string, configPath string, cmd *cobra.Command, debug bool) (bool, error) {
 	var tty io.Reader
 	isTty := true
